fix integer truncation in getTimeInTicks

+++ redeem/path_planner/Path.h
@@ -273,7 +273,7 @@ public:
     return moveMask & 255;
   }
 
-  inline unsigned long getTimeInTicks() {
+  inline unsigned long long getTimeInTicks() {
     return timeInTicks;
   }
